회원 탈퇴

+++ backend/src/main/java/com/example/backend/service/member/MemberService.java
@@ -31,7 +31,17 @@ public class MemberService {
         return mapper.selectById(id);
     }
 
-    public void remove(Member member) {
-        mapper.deleteById(member.getId());
+    public boolean remove(Member member) {
+        int cnt = 0;
+
+        //기존 암호와 비교
+        Member db = mapper.selectById(member.getId());
+
+        if (db != null) {
+            if (db.getPassword().equals(member.getPassword())) {
+                cnt = mapper.deleteById(member.getId());
+            }
+        }
+        return cnt == 1;
     }
 }
